fixed bug in MBUtils, stripBlankEnds where lines such as the following: "\t \r"
would core dump.  Checking the string the front only stripped characters
that are tab and space leaving and index of 2 while checking the string
from the end also stripped out carriage returns and newlines leaving
an index of -1.  This case was not checked and the the calculated length
(end - start) + 1 = -2 would cause "new char[length]" to throw an exception



git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@1768 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf