feat(ci): Added automated builds for main.
Added automated builds for the main branch and pull requests to the
main branch using GitHub Workflows. This will build on both Ubuntu
and Mac OS X. In order to get the buidls to fail correctly, some
of the build scripts needed to be updated to return an error code
if the build failed.

Fixes: #3