Contributor should have right to enter agreement
Using the lite workflow, there isn't a trivial way to mark individuals as having accepted the CLA under an employer.

So, if an employer has entered under an agreement, and authorized their employee, this bit should allow the employee to comment the acceptance phrase.

We're changing the language of the template here, so be careful.