added nspatch, mods to nsplug, for better autotesting (#61)
New app nspatch: ingests (1) a "stem" moos or bhv file, (2) one or more
patch moos or bhv files, and outputs a modified moos file with the
patch(es) applied to the stem. Designed to be a core tool for generating
auto-test missions with only slight test/evaluation conditions, from a
single stem mission. The older sister app, nsplug, was augmented with
the -x or --xfile flag, to first look for a variant of the input file,
e.g., meta_vehicle.moos, with the name meta_vehicle.moosx. The latter
will at times be the output of an nspatch stage.