Adding a parameter so that a string can be appended on the end of the names of the three MOOS variables published by the PID controller that are associated with control values.
Example: The parameter value of _PID will cause the app to publish
the value for rudder to DESIRED_RUDDER_PID, and the same behavior
for the other thrust and elevator variables.

The value read in the config block is transformed to upper case.
The default value is empty, or nothing is appended to the end.
The default behavior of this app is unchanged.