Run CI on 24.8.1 release branch
We typically only run CI & release from master.
However, we're making a backport branch and releasing 24.8.1.

To get CI rolling, since this branch won't be merged, manually include
it in the list.
If we expect more backport releases, we should probably swap to run CI
on any branches matching a pattern (or all branches).